+++ service/controller/resource/endpoint/create.go
@@ -7,7 +7,6 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/types"
 
 	"github.com/giantswarm/kvm-operator/service/controller/key"
 )
@@ -40,22 +39,32 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 		r.logger.Debugf(ctx, "determined node pod is not ready")
 	}
 
+	var service corev1.Service
+	{
+		err := r.ctrlClient.Get(ctx, serviceName, &service)
+		if err != nil {
+			return microerror.Mask(err)
+		}
+	}
+
 	var needCreate bool
 	var endpoints corev1.Endpoints
 	{
 		r.logger.Debugf(ctx, "retrieving endpoints %#q", serviceName)
 		err = r.ctrlClient.Get(ctx, serviceName, &endpoints)
 		if errors.IsNotFound(err) {
+			r.logger.Debugf(ctx, "endpoints not found, creating")
 			needCreate = true
 		} else if err != nil {
 			r.logger.Debugf(ctx, "error retrieving endpoints")
 			return microerror.Mask(err)
+		} else {
+			r.logger.Debugf(ctx, "retrieved endpoints")
 		}
-		r.logger.Debugf(ctx, "retrieved endpoints")
 	}
 
 	if needCreate {
-		err := r.createEndpoints(ctx, serviceName, readyForTraffic, nodeIP)
+		err := r.createEndpoints(ctx, service, readyForTraffic, nodeIP)
 		if err != nil {
 			return microerror.Mask(err)
 		}
@@ -69,13 +78,7 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 	return nil
 }
 
-func (r *Resource) createEndpoints(ctx context.Context, serviceName types.NamespacedName, readyForTraffic bool, nodeIP string) error {
-	var service corev1.Service
-	err := r.ctrlClient.Get(ctx, serviceName, &service)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
+func (r *Resource) createEndpoints(ctx context.Context, service corev1.Service, readyForTraffic bool, nodeIP string) error {
 	var endpoints corev1.Endpoints
 	{
 		var ports []corev1.EndpointPort
@@ -99,8 +102,8 @@ func (r *Resource) createEndpoints(ctx context.Context, serviceName types.Namesp
 
 		endpoints = corev1.Endpoints{
 			ObjectMeta: v1.ObjectMeta{
-				Name:      serviceName.Name,
-				Namespace: serviceName.Namespace,
+				Name:      service.Name,
+				Namespace: service.Namespace,
 			},
 			Subsets: []corev1.EndpointSubset{
 				{
@@ -112,7 +115,7 @@ func (r *Resource) createEndpoints(ctx context.Context, serviceName types.Namesp
 		}
 	}
 
-	err = r.ctrlClient.Create(ctx, &endpoints)
+	err := r.ctrlClient.Create(ctx, &endpoints)
 	if err != nil {
 		return microerror.Mask(err)
 	}