+++ service/controller/resource/namespace/current.go
@@ -53,7 +53,7 @@ func (r *Resource) GetCurrentState(ctx context.Context, obj interface{}) (interf
 		return nil, nil
 	}
 
-	if namespace == nil && key.IsDeleted(customObject) {
+	if namespace == nil && key.IsDeleted(&customObject) {
 		r.logger.Debugf(ctx, "resource deletion completed")
 
 		r.logger.Debugf(ctx, "canceling resource")
@@ -69,7 +69,7 @@ func (r *Resource) GetCurrentState(ctx context.Context, obj interface{}) (interf
 	// resources in it. As soon as the draining was done and the pods got removed
 	// we get an empty list here after the delete event got replayed. Then we just
 	// remove the namespace as usual.
-	if key.IsDeleted(customObject) {
+	if key.IsDeleted(&customObject) {
 		n := key.ClusterNamespace(customObject)
 		list, err := r.k8sClient.CoreV1().Pods(n).List(ctx, metav1.ListOptions{})
 		if err != nil {